chore(server-split): v0.4.x PR-B — LLM/Ollama routes moved to routes/llm.py (12 endpoints + 5 helpers)
docs/design/v0.4.x-server-split.md 8-PR sequence 중 세 번째 PR.
PR-A.0 + PR-A.1 위에 LLM 도메인 분리. URL byte-identical 유지.

## 이동된 surface (467 lines 서버에서 제거)

### 12 endpoints
- /llm/modes/, /llm/install/, /llm/active
- /admin/llm/install-progress, /admin/llm/installed, /admin/llm/resolution
- /admin/llm/recommend, /admin/llm/pull, /admin/llm/delete
- /admin/llm/selections, /admin/llm/select (POST + DELETE)

### 5 module-level helpers
- `_model_catalog()` — chat/coding/meta/auto/wiki_edit 모드별 후보 모델
- `_allowed_install_models()` — /llm/install/ 허용 모델 set
- `_install_progress: dict` + `_install_lock` — 백그라운드 install progress 트래커
- `_start_install_with_progress(model)` — 백그라운드 thread 로 Ollama pull
- `_list_installed_ollama_models()` — ollama list HTTP

### Back-compat re-exports (server_llmwiki.py)

테스트가 `srv.<name>` 또는 `from server_llmwiki import <name>` 으로
attribute 접근 — back-compat 위해 server 에서 routes/llm 으로부터
re-import (#noqa: F401):

```python
from routes.llm import (
    _allowed_install_models,
    _install_progress,
    _model_catalog,
    _start_install_with_progress,
    llm_active,
)
```

## 변경 패턴 (handler 본문 byte-identical)

각 handler 본문 100% 보존. `@app.<m>` → `@router.<m>` 만 변경.
모듈 helpers 도 그대로 (move 만).

## server_llmwiki.py 변경

- 12 endpoint + 5 helper 제거 (~467 lines)
- `from routes.llm import router as llm_router` + 5 back-compat re-export
- `app.include_router(llm_router)` (auth_router 다음)

## 테스트 인프라 — tests/_server_split_helpers.py 신설

source-level 스캔하는 테스트 6 개가 routes/llm.py 도 봐야 함.
`combined_server_source()` 헬퍼: server + routes/* 소스 concat +
`@router.` → `@app.` 정규화. 향후 PR-C..PR-H 가 새 routes/<domain>.py
추가해도 헬퍼만 갱신하면 됨 (테스트 변경 0).

업데이트된 테스트:
- tests/test_chat_mode_picker.py (2 setUpClass)
- tests/test_install_progress.py (3 setUpClass)
- tests/test_model_catalog_per_mode.py (2 setUpClass)
- tests/test_model_names_install.py (2 setUpClass)
- tests/test_model_resolver.py (1 setUpClass)
- tests/test_a2_phase2_selected_model.py (1 setUpClass — concat 패턴 확장)
- tests/test_llm_active_endpoint.py — `patch.object(srv, ...)` → `patch("routes.llm.verify_api_key", ...)` (모듈 binding 변경 반영)

## 검증 (design memo §5 contract)

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ `python scripts/audit_endpoint_paths.py origin/main` → **OK: 125 endpoints identical to origin/main** |
| 2. Middleware order | ✓ test_middleware_stack_includes_rate_limit_and_no_cache pass |
| 3. _write_audit emit | ✓ 24 사이트 동일 (routes/_helpers SOT) |
| 4. /static mount | ✓ test_static_mount_present pass |
| 5. startup event | ✓ test_startup_event_registered pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** (2 .env-flakes deselected — pre-existing) |
| ruff F-class | ✓ All checks passed |

## 변경 통계

```
 server_llmwiki.py                      | 484 +--------------------------------
 tests/test_a2_phase2_selected_model.py |  10 +-
 tests/test_chat_mode_picker.py         |   8 +-
 tests/test_install_progress.py         |  11 +-
 tests/test_llm_active_endpoint.py      |   7 +-
 tests/test_model_catalog_per_mode.py   |   7 +-
 tests/test_model_names_install.py      |   8 +-
 tests/test_model_resolver.py           |   5 +-
 8 files changed, 47 insertions(+), 493 deletions(-)
 (+) routes/llm.py:                       520 lines (new)
 (+) tests/_server_split_helpers.py:       48 lines (new)
```

server_llmwiki.py: 5170 → 4685 lines (-485). 8-PR roadmap 진행률
≈ **20%** ((724+485)/5894).

## Out of scope (다음 PR 들로)

- PR-C: /jobs/* (W8 workspace jobs)
- PR-D: /artifacts/* + /admin/artifacts/* (W7-A data artifacts)
- PR-E: Phase 7 자기진화 + 멀티모달
- PR-F: /code/* (코딩 에이전트)
- PR-G: 잔여 /admin/*
- PR-H: server_llmwiki.py 최종 정리 (~500 lines target)

Quality delta: exempt (label: chore). URL 표면 byte-identical, core/ 무관.
